Add Constants class to hold enums and Ui to deal with ui stuff
